refactor(gamma-apim): split api types into domain files
