+++ gravitee-gamma/gravitee-gamma-module-apim/src/main/ui/features/apis/constants/alertConstants.ts
@@ -22,7 +22,7 @@ import type {
     AlertRuleId,
     AlertStringOperator,
     AlertTimeUnit,
-} from '../types/api';
+} from '../types';
 
 // ─── Rule definitions ──────────────────────────────────────────────────────────
