+++ gravitee-gamma/gravitee-gamma-module-apim/src/main/ui/features/apis/pages/detail/alerts/FilterRow.tsx
@@ -36,7 +36,7 @@ import {
     isStringMetric,
     type AlertMetricDefinition,
 } from '../../../constants/alertConstants';
-import type { AlertConditionType, AlertFormCondition, AlertOperator, AlertStringOperator } from '../../../types/api';
+import type { AlertConditionType, AlertFormCondition, AlertOperator, AlertStringOperator } from '../../../types';
 
 interface Props {
     filter: AlertFormCondition;
