+++ gravitee-gamma/gravitee-gamma-module-apim/src/main/ui/features/apis/pages/detail/alerts/HistoryTab.tsx
@@ -27,7 +27,7 @@ import {
     TableRow,
 } from '@gravitee/graphene-core';
 
-import type { AlertHistoryPage } from '../../../types/api';
+import type { AlertHistoryPage } from '../../../types';
 
 export interface HistoryTabProps {
     historyPage: AlertHistoryPage | undefined;
