+++ gravitee-gamma/gravitee-gamma-module-apim/src/main/ui/features/apis/pages/detail/alerts/RateConditionForm.tsx
@@ -22,7 +22,7 @@ import {
     isStringMetric,
     type AlertMetricDefinition,
 } from '../../../constants/alertConstants';
-import type { AlertFormCondition, AlertOperator, AlertStringOperator, AlertTimeUnit } from '../../../types/api';
+import type { AlertFormCondition, AlertOperator, AlertStringOperator, AlertTimeUnit } from '../../../types';
 
 interface Props {
     condition: AlertFormCondition;
