+++ gravitee-gamma/gravitee-gamma-module-apim/src/main/ui/features/apis/pages/detail/alerts/SimpleConditionForm.tsx
@@ -21,7 +21,7 @@ import {
     getConditionTypesForMetric,
     type AlertMetricDefinition,
 } from '../../../constants/alertConstants';
-import type { AlertConditionType, AlertFormCondition, AlertOperator, AlertStringOperator } from '../../../types/api';
+import type { AlertConditionType, AlertFormCondition, AlertOperator, AlertStringOperator } from '../../../types';
 
 interface Props {
     condition: AlertFormCondition;
