+++ gravitee-gamma/gravitee-gamma-module-apim/src/main/ui/features/apis/pages/detail/alerts/useAlertForm.ts
@@ -44,7 +44,7 @@ import type {
     AlertNotificationChannel,
     AlertRuleId,
     AlertSeverity,
-} from '../../../types/api';
+} from '../../../types';
 import { apiAlertKeys } from '../../../utils/queryKeys';
 
 function getDefaultCondition(ruleId: AlertRuleId): AlertFormCondition[] {
