+++ gravitee-gamma/gravitee-gamma-module-apim/src/main/ui/features/apis/pages/detail/deployment/SingleEventDialog.tsx
@@ -27,7 +27,7 @@ import { XIcon } from '@gravitee/graphene-core/icons';
 import { useMemo } from 'react';
 
 import { extractDefinition, formatDate } from './utils';
-import type { ApiEvent } from '../../../types/api';
+import type { ApiEvent } from '../../../types';
 
 type Props = Readonly<{
     event: ApiEvent;
