+++ gravitee-gamma/gravitee-gamma-module-apim/src/main/ui/features/apis/pages/detail/properties/ApiPropertiesPage.tsx
@@ -65,7 +65,7 @@ import { useParams } from 'react-router-dom';
 
 import { useApiDetail } from '../../../hooks/useApiDetail';
 import { updateApiProperties } from '../../../services/apis';
-import type { Property } from '../../../types/api';
+import type { Property } from '../../../types';
 import { parsePropertiesStringFormat } from '../../../utils/propertiesParser';
 import { apiDetailKeys } from '../../../utils/queryKeys';
