+++ gravitee-gamma/gravitee-gamma-module-apim/src/main/ui/features/apis/services/alerts.ts
@@ -26,7 +26,7 @@ import type {
     AlertOperator,
     AlertStringOperator,
     AlertTrigger,
-} from '../types/api';
+} from '../types';
 
 // ─── Condition converters ──────────────────────────────────────────────────────
