+++ gravitee-gamma/gravitee-gamma-module-apim/src/main/ui/features/apis/types/index.ts
@@ -13,9 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { apimFetchJsonV1Env } from '../../../shared/api/apimClient';
-import type { EnvCategory } from '../types';
 
-export async function getEnvCategories(environmentId: string): Promise<EnvCategory[]> {
-    return apimFetchJsonV1Env<EnvCategory[]>(environmentId, '/configuration/categories');
-}
+export * from './alert';
+export * from './analytics';
+export * from './api';
+export * from './apiCreation';
+export * from './auditLogs.types';
+export * from './broadcast';
+export * from './members.types';
+export * from './notification';
+export * from './plan';
+export * from './subscription';
